Some refactoring done like beam id variable type changes to uint32_t (was uint16_t). Installation method of beam helper divided smaller part (several methods).

+++ helper/satellite-geo-helper.cc
@@ -22,7 +22,6 @@
 #include "ns3/packet.h"
 #include "ns3/names.h"
 #include "ns3/uinteger.h"
-//#include "../model/satellite-channel.h"
 #include "../model/satellite-geo-net-device.h"
 #include "../model/satellite-phy.h"
 #include "../model/satellite-phy-tx.h"
@@ -105,7 +104,7 @@ SatGeoHelper::Install (std::string nName)
 }
 
 void
-SatGeoHelper::AttachChannels (Ptr<NetDevice> d, Ptr<SatChannel> ff, Ptr<SatChannel> fr, Ptr<SatChannel> uf, Ptr<SatChannel> ur, uint16_t beamId )
+SatGeoHelper::AttachChannels (Ptr<NetDevice> d, Ptr<SatChannel> ff, Ptr<SatChannel> fr, Ptr<SatChannel> uf, Ptr<SatChannel> ur, uint32_t beamId )
 {
   NS_LOG_FUNCTION (this << ff << fr << uf << ur);
 
@@ -125,7 +124,7 @@ SatGeoHelper::AttachChannels (Ptr<NetDevice> d, Ptr<SatChannel> ff, Ptr<SatChann
   // Configure the SatPhyRxCarrier instances
   // Note, that these have to be changed so that they match the real frame configuration.
   // Now we survive with one SatPhyRxCarrier, because we do not have a NCC scheduler.
-  uint16_t RETURN_CARRIERS(1);
+  uint32_t RETURN_CARRIERS(1);
   uPhyRx->ConfigurePhyRxCarriers (RETURN_CARRIERS);
 
   fPhyTx->SetChannel (fr);
@@ -133,7 +132,7 @@ SatGeoHelper::AttachChannels (Ptr<NetDevice> d, Ptr<SatChannel> ff, Ptr<SatChann
   fPhyRx->SetDevice (dev);
 
   // By default, there is one carrier in the forward link
-  uint16_t FORWARD_CARRIERS(1);
+  uint32_t FORWARD_CARRIERS(1);
   fPhyRx->ConfigurePhyRxCarriers (FORWARD_CARRIERS);
 
   SatPhy::ReceiveCallback uCb = MakeCallback (&SatGeoNetDevice::ReceiveUser, dev);
